cleanup modules a bit
